django-app/requirements.txt
asgiref==3.11.0
attrs==25.4.0
Django==5.2.8
django-cleanup==9.0.0
django-cors-headers==4.9.0
django-js-asset==3.1.2
django-mptt==0.18.0
django-mptt-admin==2.8.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
drf-spectacular==0.29.0
easy-thumbnails==2.10.1
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
packaging==25.0
pillow==12.0.0
pipdeptree==2.30.0
psycopg==3.2.13
PyJWT==2.10.1
python-decouple==3.8
PyYAML==6.0.3
referencing==0.37.0
rpds-py==0.29.0
sqlparse==0.5.3
typing_extensions==4.15.0
uritemplate==4.2.0
